rename session getbyuserid to getbyactiveuserid
