+++ packages/schema/src/accounts/session/getByActiveUserId.ts
