+++ packages/schema/src/accounts/session/index.ts
@@ -1,2 +1,2 @@
-export * as getByUserId from './getByUserId';
+export * as getByActiveUserId from './getByActiveUserId';
 export * as deleteById from './deleteById';
