+++ packages/server/src/accounts/Session.ts
@@ -6,9 +6,9 @@ import {Action, UserData} from '../types';
 export function Session(em: EntityManager): Action[] {
 	return [
 		{
-			route: Schema.accounts.session.getByUserId.route,
-			action: async (userData: UserData): Promise<Schema.accounts.session.getByUserId.response> => {
-				const sessions = await new controllers.sessionController(em, userData).getByUserId();
+			route: Schema.accounts.session.getByActiveUserId.route,
+			action: async (userData: UserData): Promise<Schema.accounts.session.getByActiveUserId.response> => {
+				const sessions = await new controllers.sessionController(em, userData).getByActiveUserId();
 
 				// @ts-ignore
 				return sessions.map(session => {
